requirements_production.txt
# Core Dependencies
moviepy==1.0.3
librosa==0.10.0
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0

# Arabic/Text Processing
arabic-reshaper==0.6.0
python-bidi==0.4.2

# Streamlit (Web UI)
streamlit==1.28.1

# FastAPI (Backend API)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Production Deployment
gunicorn==21.2.0
python-dotenv==1.0.0

# Development/Debugging
ipython==8.16.0
